02_Python/04_WebScrapper/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
psycopg==3.1.18
psycopg-binary==3.1.18
requests==2.31.0
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
